NAYA-120 Justtype not fullscreen on topaz
Make sure to also resize the ipc buffer to the correct orientation
through a flip event when justtype is added to the overlay
windowmanager. When SysMgr is shown in portrait mode first and
only rotated after unlocking, the ipc buffer belonging to the
JustType app is not flipped correctly, causing the dimensions in
WindowedWebApp to get out of sync.

Open-webOS-DCO-1.0-Signed-off-by: Florian Haenel <florian.hanel@palm.com>

Change-Id: I3b6ddc34a3437098ecaa7fe99039d7954b3ffc23