LockWindow::requiresPasscode check policy before returning true
Previous behavior checked only if an EAS Policy was pending. New
behavior checks that and also checks if a password is required in the
Policy.  By default it is not.

This should eliminate the annoyance of being required to set a PIN
on unlock when it is not in fact required.

Take fix for LockWindow::requiresPasscode [OWEBOS-2735]

Open-webOS-DCO-1.0-Signed-off-by: Eric Blade <eric.blade@palm.com>

Change-Id: I380601039f0f3687c296105de7f967afc0dea571