Fix system menu not showing
At the beginning of the Qt 4.8 -> Qt 5 transition the QtDeclarative
module (QML 1) was not a part of the desktop build, so references to
it were replaced with ones to the new QtQuick (QML 2) module, but
the QtDeclarative module has been added to the build afterwards.

Using the QtQuick module causes old QML 1 stuff such as the system
menu to not be displayed. Using the QtDeclarative module fixes this.

Open-webOS-DCO-1.0-Signed-off-by: Teppo Pennanen <teppo.pennanen@palm.com>

Change-Id: I4a1fa35a4084caa99456f57283a50ffa7d650ccc