Return and tab events generating wrong QKeyEvent text
On generating a key event for Qt::Key_Return, Qt::Key_Enter and
Qt::Key_Tab on IME, it was using as text a  direct conversion of
the Qt::Key to a QChar. This is wrong for these characters.

This patch maps explicitely these characters to \r and \n.

Open-webOS-DCO-1.0-Signed-off-by: Jose Dapena Paz <jose.dapena@palm.com>

Change-Id: I3796ad30450f9c6f3fdaa653026b0531b1826f90