Fix MouseEventEater check for mouse position
The MouseEventEater checks if the event is in the scene, and drops
the event if not. But the check is wrong as it is using globalPos
(position in screen) against positions related to scene.

This makes events fail on desktop depending on the position of the
emulator window.
Open-webOS-DCO-1.0-Signed-off-by: Jose Dapena Paz <jose.dapena@palm.com>

Change-Id: I8699490874baf73d12d4e5d51eb24701b9423bb4
Reviewed-on: https://g2g.palm.com/1649
Reviewed-by: Build Verification
Reviewed-by: Jose Dapena <jose.dapena@palm.com>
Tested-by: Jose Dapena <jose.dapena@palm.com>
Reviewed-by: Teppo Pennanen <teppo.pennanen@palm.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>